fixed gathering skill data from API

+++ Assets/Scripts/API/Response/RiotAPIChampionDataResponse.cs
@@ -55,7 +55,7 @@ public class Leveling
 
 public class Modifier
 {
-    public List<double> values { get; set; }
+    public List<float> values { get; set; }
     public List<string> units { get; set; }
 }
